refactor(desktop): split skills store
